+++ src/ATen/native/xpu/sycl/HermitePolynomialHKernel.cpp
@@ -1,13 +1,13 @@
 #include <ATen/Dispatch.h>
 #include <ATen/native/BinaryOps.h>
 #include <ATen/native/Math.h>
-#include <ATen/native/xpu/sycl/Hermite_polynomial_he.h>
+#include <ATen/native/xpu/sycl/HermitePolynomialHKernel.h>
 #include <ATen/native/xpu/sycl/Loops.h>
 
 namespace at::native::xpu {
 
 template <typename scalar_t>
-struct hermite_polynomial_he_functor {
+struct hermite_polynomial_h_functor {
   scalar_t operator()(scalar_t x, scalar_t n_) const {
     int64_t n = static_cast<int64_t>(n_);
     if (n < 0) {
@@ -19,15 +19,15 @@ struct hermite_polynomial_he_functor {
     }
 
     if (n == 1) {
-      return x;
+      return x + x;
     }
 
     scalar_t p = scalar_t(1.0);
-    scalar_t q = x;
-    scalar_t r;
+    scalar_t q = x + x;
+    scalar_t r = scalar_t(0.0);
 
-    for (int64_t k = 1; k < n; k++) {
-      r = x * q - k * p;
+    for (int64_t k = 2; k < n + n; k += 2) {
+      r = (x + x) * q - k * p;
       p = q;
       q = r;
     }
@@ -36,11 +36,11 @@ struct hermite_polynomial_he_functor {
   }
 };
 
-void hermite_polynomial_he_kernel(TensorIteratorBase& iterator) {
+void hermite_polynomial_h_kernel(TensorIteratorBase& iterator) {
   AT_DISPATCH_FLOATING_TYPES(
-      iterator.common_dtype(), "hermite_polynomial_he_xpu", [&]() {
+      iterator.common_dtype(), "hermite_polynomial_h_xpu", [&]() {
         gpu_kernel_with_scalars(
-            iterator, hermite_polynomial_he_functor<scalar_t>());
+            iterator, hermite_polynomial_h_functor<scalar_t>());
       });
 }
