+++ src/ATen/native/xpu/sycl/HermitePolynomialHKernel.h
