+++ src/ATen/native/xpu/sycl/HermitePolynomialHeKernel.h
